added CLI option composeform for generating a message based on Winlink forms, like ICS213
