+++ README.md
@@ -68,6 +68,7 @@ Copyright (c) 2020 Martin Hebnes Pedersen LA5NTA
 * K0SWE - Chris Keller
 * KD8DRX - Will Davidson
 * KE8HMG - Andrew Huebner
+* KI7RMJ - Rainer Grosskopf
 * LA3QMA - Kai Günter Brandt
 * LA4TTA - Erlend Grimseid
 * LA5NTA - Martin Hebnes Pedersen
