+++ man/pat.1
@@ -17,6 +17,9 @@ Run http server for web gui.
 \fIcompose\fP
 Compose a new message.
 .TP
+\fIcomposeform\fP
+Compose a new message based on a Winlink-style form, e.g. the ICS213 form.
+.TP
 \fIread\fP
 Read Messages.
 .TP
